refactored HandEvaluator from commandline to its own class
